creating hocs for different radar components
