+++ src/sections/PerformanceRadar/index.js
@@ -1,1 +1,4 @@
-export { default as PerformanceRadar } from './Tutela';
+import Tutela from './Tutela';
+import Pip from './PIP';
+
+export default { Tutela, Pip };
