+++ src/sections/index.js
@@ -1,5 +1,5 @@
 export { default as Alerts } from './Alerts';
-export * from './PerformanceRadar';
+export { default as PerformanceRadar } from './PerformanceRadar';
 export { default as ProcessingTime } from './ProcessingTime';
 export { default as ProcessList } from './ProcessList';
 export { default as Today } from './Today';
